auth/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.6.0
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==4.2
django-cors-headers==3.14.0
django-import-export==4.0.9
django-livereload-server==0.4
django-nextjs==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-extra-fields==3.7.0
filetype==1.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
PyJWT==2.6.0
pytz==2023.3
requests==2.32.3
setuptools==70.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
tornado==6.3.1
tzdata==2023.3
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
